requirements.txt
kazoo
pexpect
matplotlib
numpy
h5py
requests
gitpython
